Updated PlaceItManager to use Long for ID

+++ src/com/ucsd/cs110w/group16/placeits/PlaceItManager.java
@@ -49,7 +49,7 @@ public class PlaceItManager
 		return list;
 	}
 	
-	public PlaceIt getPlaceIt(int ID)
+	public PlaceIt getPlaceIt(long ID)
 	{
 		for (int i = 0; i < placeits.size(); i++)
 		{
